requirements.txt
appdirs==1.4.3
bcrypt==3.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
flash==1.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.16.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.1
packaging==20.3
pandas==1.2.4
pdfminer.six==20201018
pep517==0.8.2
Pillow==8.2.0
progress==1.5
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
regex==2021.3.17
requests==2.22.0
retrying==1.3.3
sacremoses==0.0.44
sentencepiece==0.1.91
six==1.14.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SQLAlchemy==1.4.5
tokenizers==0.8.1rc1
torch==1.4.0
tqdm==4.59.0
transformers==3.0.2
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
WTForms==2.3.3
